==> pipeline/step0_load_data_in_csv.ipynb <==
# Import libraries

# http://datalytics.ru/all/kak-ispolzovat-google-bigquery-s-pomoschyu-python/
import pandas as pd
import os
import pandas_profiling
os.environ["GOOGLE_APPLICATION_CREDENTIALS"] ='data/dev-rep-final-projects-326152cab3fb.json'

# !pip install --upgrade google-cloud-bigquery[pandas]

%load_ext google.cloud.bigquery

from google.cloud import bigquery
client = bigquery.Client()

# Test connections
sql = """
SELECT *
FROM `dev-rep-final-projects.au_weather.location_labels` AS location_labels
LIMIT 15
"""

df = client.query(sql).to_dataframe()
print(df.shape)
df.head()

# main query


sql = """
SELECT  cloud_info.Date,
        cloud_info.Location_ID,
        location_labels.Location,
        cloud_info.Cloud9am,
        cloud_info.Cloud3pm,
        humidity_info.Humidity9am,
        humidity_info.Humidity3pm,
        pressure_info.Pressure9am,
        pressure_info.Pressure3pm,
        rain_info.RISK_MM,
        rain_info.RainTomorrow,
        temperature_info.MinTemp,
        temperature_info.MaxTemp,
        temperature_info.Temp9am,
        temperature_info.Temp3pm,
        weather_info.Rainfall,
        weather_info.Evaporation,
        weather_info.Sunshine,
        weather_info.RainToday,
        wind_info.WindGustDir,
        wind_info.WindGustSpeed,
        wind_info.WindDir9am,
        wind_info.WindDir3pm,
        wind_info.WindSpeed9am,
        wind_info.WindSpeed3pm,
        
humidity_info.Date AS Data2, humidity_info.Location_ID AS Location_ID2
FROM `dev-rep-final-projects.au_weather.cloud_info` AS cloud_info
LEFT JOIN `dev-rep-final-projects.au_weather.humidity_info` AS humidity_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.pressure_info` AS pressure_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.rain_info` AS rain_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.temperature_info` AS temperature_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.weather_info` AS weather_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.wind_info` AS wind_info
USING(Date, Location_ID)
LEFT JOIN `dev-rep-final-projects.au_weather.location_labels` AS location_labels
USING(Location_ID)



"""
# ON cloud_info.Date = humidity_info.Date
df = client.query(sql).to_dataframe()
print(df.shape)
df.head()

df.to_csv('data/weather.csv', sep=',')

# %%bigquery
# SELECT cloud.Date, humidity.Date AS Data2
# FROM `dev-rep-final-projects.au_weather.cloud_info` AS cloud
# LEFT JOIN `dev-rep-final-projects.au_weather.humidity_info` AS humidity
# ON cloud.Date = humidity.Date
# LIMIT 5;

# pandas_profiling

profile = df.profile_report(title='Pandas Profiling Weather Report')
profile.to_file(output_file="data/weather.html")
