==> pipeline/step3_stacking.ipynb <==
# Libraries

import os
import pandas as pd
import numpy as np
import scipy as sp
from tqdm import tqdm_notebook
from IPython.display import display
from datetime import datetime
import joblib
from pprint import pprint

from sklearn.model_selection import train_test_split, cross_val_score, StratifiedKFold, RandomizedSearchCV, GridSearchCV
from sklearn.compose import ColumnTransformer
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier, AdaBoostClassifier, \
            GradientBoostingClassifier, BaggingClassifier
from sklearn.svm import SVC
from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.metrics import roc_auc_score, confusion_matrix, f1_score, classification_report, \
            accuracy_score, precision_score, recall_score

from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier

SEED = 24
N_JOBS = -1
cv_number = 3
cv = StratifiedKFold(n_splits=cv_number, shuffle=True, random_state=SEED)

metric = 'roc_auc'
# https://scikit-learn.org/stable/modules/model_evaluation.html#scoring-parameter
file_name_log = 'models/df_log.pkl'
path_for_tunning_model = 'models/tunning'

# List of models

model_dict = {
    'logreg': LogisticRegression(max_iter=100, random_state=SEED, n_jobs=N_JOBS),
#     'knn': KNeighborsClassifier(n_jobs=N_JOBS),
    'dt': DecisionTreeClassifier(random_state=SEED),
    'rf': RandomForestClassifier(random_state=SEED, n_jobs=N_JOBS),
    'ext': ExtraTreesClassifier(random_state=SEED, n_jobs=N_JOBS),
#     'bag': BaggingClassifier() # Dublicate RandomForest
    'adb': AdaBoostClassifier(base_estimator=DecisionTreeClassifier(max_depth=1), random_state=SEED),
    'gb': GradientBoostingClassifier(random_state=SEED),
    'xgb': XGBClassifier(random_state=SEED, n_jobs=N_JOBS),
#     'svc': SVC(probability=True, random_state=SEED),
    'cat': CatBoostClassifier(random_state=SEED, verbose=False),
    'lgb': LGBMClassifier(random_state=SEED, n_jobs=N_JOBS),
    'mlp': MLPClassifier(hidden_layer_sizes=(128, 32, 2), random_state=SEED),
}

# Load data with all imputation

df = pd.read_csv('data/df_after_all_imputations.csv')
df.sort_values(by=['Location_ID', 'Date'], inplace=True)
df.reset_index(inplace=True);
df.drop(['Unnamed: 0', 'index', 'Date'], axis=1, inplace=True)

print(df.shape)
df.head(3).T
# df.info()


# Split data test for stacking 

X, X_test_stacking, y, y_test_stacking = train_test_split(
    df.drop('RainTomorrow', axis=1), 
    df['RainTomorrow'],
    test_size=0.2,
    random_state=SEED,
    shuffle=True,
)
print(X.shape, X_test_stacking.shape, y.shape, y_test_stacking.shape)

def create_log_file(file_name):
    d = dict(
        model_type=model_type,
        model=model,
        type_of_model='tunning', # or base
        best_preprocessing_pipe=best_preprocessing_pipe,
        time_save=datetime.now().strftime("%m-%d-%Y, %H-%M-%S"),
        time_search=None,
        best_roc_auc=0.5,
        best_hypparam=None,
        type_of_searchers='GridSearchCV',
        start_params_for_searchers=None,
    )
    df_log = pd.DataFrame(columns=d.keys())
    joblib.dump(df_log, file_name)

def write_log(data, file_name):
    if not os.path.isfile(file_name):
        create_log_file(file_name)
    df_log = joblib.load(file_name)
    df_log = pd.concat([df_log, data], ignore_index=True)
    joblib.dump(df_log, file_name)


# LOG file display

df_log0 = joblib.load(file_name_log)
list_display = ['model_type', 'type_of_model', 'best_roc_auc','time_save', 'time_search', 'best_hypparam', 'model' ]
df_log = df_log0.sort_values(by=['model_type', 'best_roc_auc'], ascending=[True, False])[list_display]
display(df_log)
df_log = df_log.groupby(by=['model_type']).agg(['first'])
display(df_log)

# get full list(dict) of tunning models
# drop stacking model
# 'model' - path to file with model
dict_tunning_models = dict(zip(df_log.index.to_list(), df_log[('model', 'first')]))
print(dict_tunning_models)
dict_tunning_models.pop('stacking', None)
# dict_tunning_models.pop('ext', None)
# dict_tunning_models.pop('rf', None)
pprint(dict_tunning_models)

# для Слайда о препроцессинге

for model_type, model_path in dict_tunning_models.items():
#     print(model_type, model_path) 
    load_model = joblib.load(dict_tunning_models[model_type])
    print(model_type, load_model.steps[0][1].steps)

# FeatureUnion for Stacking

# Xnew_train
for model_type, model_path in dict_tunning_models.items():
    print(model_type, model_path)

class PredictProbaToTransform(BaseEstimator, TransformerMixin):
    """
    Give all X and classifier on input
    output - predict_proba by classifier (only binary classifier)
    This class needs to compare outputs from classifiers in FeatureUnion model
    """
    def __init__(self, clf):
        self.clf = clf
        
    def fit(self, X, y=None):
        return self
    
    def transform(self, X):
        output = self.clf.predict_proba(X)[:, 1]
        output = np.reshape(output, (-1,1))
#         print(output.shape)
        return output


start_time = datetime.now()
union_list = []

for model_type, model_path in dict_tunning_models.items():
    if os.path.exists(model_path):
        load_model = joblib.load(dict_tunning_models[model_type])
    else:
        print(f'{model_path} - file not found')
        continue
    union_list.append((model_type, PredictProbaToTransform(load_model)))

union = FeatureUnion(union_list)

stacking_pipe = Pipeline([
    ('union', union),
    ('meta_m', LogisticRegression()) 
])

# Hold out
X_train, X_test, y_train, y_test = train_test_split(
    X_test_stacking, 
    y_test_stacking,
    test_size=0.2,
    random_state=SEED,
    shuffle=True,
)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

stacking_pipe.fit(X_train, y_train)
y_pred_train = stacking_pipe.predict_proba(X_train)[:, 1]
y_pred_test = stacking_pipe.predict_proba(X_test)[:, 1]
print(f'roc_auc train - {roc_auc_score(y_train, y_pred_train):.5f}, roc_auc test - {roc_auc_score(y_test, y_pred_test):.5f}')

stacking_pipe.fit(X_test_stacking, y_test_stacking)

model_type = 'stacking'
type_of_searchers='HoldOut'
cur_time = datetime.now().strftime("%m-%d-%Y, %H-%M-%S")
type_of_model = 'base_light'
file_name = os.path.join(path_for_tunning_model, \
    f'{model_type}-{type_of_model}_model-[{type_of_searchers}][{cur_time}].pkl')
joblib.dump(stacking_pipe, file_name)
new_log = dict(
    model_type=model_type,
    model=file_name,
    type_of_model=type_of_model,
    best_preprocessing_pipe='?',
    time_save=cur_time,
    time_search=datetime.now() - start_time,
    best_roc_auc=roc_auc_score(y_test, y_pred_test),
    best_hypparam='',
    type_of_searchers=type_of_searchers,
    start_params_for_searchers='',
)
print(f'new_log[best_roc_auc] = {new_log["best_roc_auc"]}')
write_log(pd.DataFrame([new_log]), file_name_log)

print(stacking_pipe.steps[1][1].coef_[0])
print(np.array(union.transformer_list)[:,0])
# [2.05081497 1.07581275 2.14593311 2.95698225]
# ['adb' 'gb' 'logreg' 'xgb']


write_log(pd.DataFrame([new_log]), file_name_log)

load_model = joblib.load(dict_tunning_models['logreg'])
ppp = PredictProbaToTransform(load_model)
display(ppp.fit_transform(X_test_stacking)[:5])
union_list = []

for model_type, model_path in dict_tunning_models.items():
    print(model_type, model_path)
#     to check file for existing
    if os.path.exists(model_path):
        load_model = joblib.load(dict_tunning_models[model_type])
    else:
        print(f'{model_path} - file not found')
        continue
    union_list.append((model_type, PredictProbaToTransform(load_model)))
# union_list
union = FeatureUnion(union_list)
output = union.fit_transform(X_test_stacking)
display(output[:5])
