==> pipeline/step4_django_connect.ipynb <==
# Libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm_notebook
from IPython.display import display
import time
import joblib
from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score, confusion_matrix, f1_score, classification_report, \
            accuracy_score, precision_score, recall_score
from sklearn.base import BaseEstimator, TransformerMixin

from pprint import pprint # хорошо словари принтит

file_name_log = 'models/df_log.pkl'
file_name_log2 = 'models/df_log2.pkl'
# path_for_tunning_model = 'models/tunning'
SEED = 24

class PredictProbaToTransform(BaseEstimator, TransformerMixin):
    """
    Give all X and classifier on input
    output - predict_proba by classifier (only binary classifier)
    This class needs to compare outputs from classifiers in FeatureUnion model
    """
    def __init__(self, clf):
        self.clf = clf
        
    def fit(self, X, y=None):
        return self
    
    def transform(self, X):
        output = self.clf.predict_proba(X)[:, 1]
        output = np.reshape(output, (-1,1))
        return output


# Load data (step 1)

# load data
df = pd.read_csv('data/weather.csv')
df.sort_values(by=['Location_ID', 'Date'], inplace=True)
df.reset_index(inplace=True);

print(df.shape)

# Cleaning and preparing PANDAS (step 1)

# separate target
df_target = df['RainTomorrow'].map(lambda x: 1 if x == True else 0).astype('uint8')
# load data

print(df.shape, df.duplicated().sum(), df.columns)
df.head(2)

# Template csv

# load data
df = pd.read_csv('data/weather.csv')
df.sort_values(by=['Location_ID', 'Date'], inplace=True)
df.reset_index(inplace=True);
df.drop(['Unnamed: 0', 'Data2',  'Location', 'Location_ID2', 'RISK_MM', 'RainTomorrow', 'RainToday', 'index'], axis=1, inplace=True)
print(df.shape)

template_csv = df.dropna().head()

template_csv.to_csv('data/template1.csv')

df = pd.read_csv('data/template1.csv')
df

# def preprocessing_befor_load(df):

def preprocessing_befor_load(df):
    # convert dates
    df['Date'] = pd.to_datetime(df['Date'], format='%Y-%m-%d')
    df['year'] = df['Date'].dt.year.astype('uint16')
    df['month'] = df['Date'].dt.month.astype('uint8')
    df['season'] = df['month'].apply(lambda x: 1 if x in [1,11,12] else 
                                    2 if x in [3, 4, 5] else 
                                    3 if x in [6, 7, 8] else 4).astype('uint8')
    # replace wind directions with numeric
    wind_rose = ['N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE', 'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW']
    dict_replacer = dict(zip(wind_rose, range(len(wind_rose))))
    df.replace(dict_replacer, inplace=True)

    # drop
    df.drop(['Unnamed: 0', 'Date'], axis=1, inplace=True)
    return df

print(df.shape)
df = preprocessing_befor_load(df)
print(df.shape)
display(df.head().T)

# Load data with all imputation

df = pd.read_csv('data/df_after_all_imputations.csv')
df.sort_values(by=['Location_ID', 'Date'], inplace=True)
df.reset_index(inplace=True);
df.drop(['Unnamed: 0', 'index', 'Date'], axis=1, inplace=True)

print(df.shape)

# Split data test for stacking 

X, X_test_stacking, y, y_test_stacking = train_test_split(
    df.drop('RainTomorrow', axis=1), 
    df['RainTomorrow'],
    test_size=0.2,
    random_state=SEED,
    shuffle=True,
)
print(X.shape, X_test_stacking.shape, y.shape, y_test_stacking.shape)

# LOG file display (step 3)

df_log0 = joblib.load(file_name_log)
display(df_log.tail().T)

df_log0 = joblib.load(file_name_log)
# print(df_log.columns)
display(df_log.tail().T)
list_display = ['model_type', 'type_of_model', 'best_roc_auc','time_save', 'time_search', 'best_hypparam', 'model' ]
df_log = df_log0.sort_values(by=['model_type', 'best_roc_auc'], ascending=[True, False])[list_display]
display(df_log)
df_log = df_log.groupby(by=['model_type']).agg(['first'])
display(df_log)

print(df_log.columns)
df_log.columns = [x[0] for x in df_log.columns]
print(df_log.columns)

dict_tunning_models = dict(zip(df_log.index.to_list(), df_log['model']))
for m, m_path in dict_tunning_models.items():
    print(m, m_path)

def m_predict(model_path):
    model = joblib.load(model_path)
    y_pred_train = model.predict_proba(X)[:, 1]
    y_pred_test = model.predict_proba(X_test_stacking)[:, 1]
#     print(f'roc_auc train - {roc_auc_score(y, y_pred_train):.5f}, roc_auc test - {roc_auc_score(y_test_stacking, y_pred_test):.5f}')
    print(model_path)
    return roc_auc_score(y, y_pred_train), roc_auc_score(y_test_stacking, y_pred_test)

# df_log['roc_auc_test'] = df_log['model'].apply(lambda x: m_predict(x))

# Test all model (exept stacking) on X_test_stacking Data

# file_name_log2 = 'models/df_log2.pkl'
df_log2 = joblib.load(file_name_log2)
display(df_log2.sort_values(by=['roc_auc_test'], ascending=False))
for_saving = df_log2.sort_values(by=['roc_auc_test'], ascending=False)
for_saving.to_csv('raiting_models')

# For SLIDES

tf_only_scale = StandardScaler()
tf_cat_and_num_list = [
    ('num', StandardScaler(), num_col),
    ('cat', OneHotEncoder(), cat_col),
]
tf_cat_and_num = ColumnTransformer(transformers=tf_cat_and_num_list)

pipe_preprocessing = Pipeline([
    ('tf1', tf_only_scale),
    ('tf2', tf_cat_and_num),
])

ml_pipe = Pipeline([
    ('pre', pipe_preprocessing),
    ('model', model),
])
param_gs = [{'pre__tf1': [None],},
            {'pre__tf2': [None],},
            {'pre__tf1': [None],'pre__tf2': [None],}]

gs = GridSearchCV(
    estimator=ml_pipe,
    param_grid=param_gs,
    cv=cv,
    scoring=metric,
)

model_dict = {
    'logreg': LogisticRegression(),
    'knn'   : KNeighborsClassifier(),
    'rf'    : RandomForestClassifier(),
    'ext'   : ExtraTreesClassifier(),
    'adb'   : AdaBoostClassifier(),
    'gb'    : GradientBoostingClassifier(),
    'xgb'   : XGBClassifier(),
    'cat'   : CatBoostClassifier(),
    'lgb'   : LGBMClassifier(),
    'mlp'   : MLPClassifier(hidden_layer_sizes=(128, 32, 2)),
}
logreg [('tf1', None), ('tf2', ColumnTransformer(...))]
knn    [('tf1', StandardScaler()), ('tf2', None)]
rf     [('tf1', None), ('tf2', None)]
ext    [('tf1', None), ('tf2', None)]
adb    [('tf1', StandardScaler()), ('tf2', None)]
gb     [('tf1', None), ('tf2', None)]
xgb    [('tf1', StandardScaler()), ('tf2', None)]
cat    [('tf1', None), ('tf2', ColumnTransformer(...))]
lgb    [('tf1', None), ('tf2', ColumnTransformer(...))]
mlp    [('tf1', StandardScaler()), ('tf2', None)]
